+++ src/test/groovy/ru/vyarus/dropwizard/guice/config/debug/DiagnosticBundleTest.groovy
@@ -4,6 +4,7 @@ import io.dropwizard.Application
 import io.dropwizard.Configuration
 import io.dropwizard.setup.Bootstrap
 import io.dropwizard.setup.Environment
+import ru.vyarus.dropwizard.guice.AbstractTest
 import ru.vyarus.dropwizard.guice.GuiceBundle
 import ru.vyarus.dropwizard.guice.bundle.lookup.VoidBundleLookup
 import ru.vyarus.dropwizard.guice.diagnostic.support.bundle.FooBundle
@@ -13,6 +14,7 @@ import ru.vyarus.dropwizard.guice.module.context.debug.DiagnosticBundle
 import ru.vyarus.dropwizard.guice.module.context.debug.diagnostic.DiagnosticConfig
 import ru.vyarus.dropwizard.guice.module.installer.feature.LifeCycleInstaller
 import ru.vyarus.dropwizard.guice.support.util.GuiceRestrictedConfigBundle
+import ru.vyarus.dropwizard.guice.test.spock.UseDropwizardApp
 import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyApp
 import spock.lang.Specification
 
@@ -20,8 +22,8 @@ import spock.lang.Specification
  * @author Vyacheslav Rusakov
  * @since 12.07.2016
  */
-@UseGuiceyApp(App)
-class DiagnosticBundleTest extends Specification {
+@UseDropwizardApp(App) // important to track HK part also
+class DiagnosticBundleTest extends AbstractTest {
 
     def "Check logging"() {
 
@@ -46,12 +48,9 @@ class DiagnosticBundleTest extends Specification {
         void initialize(Bootstrap<Configuration> bootstrap) {
             bootstrap.addBundle(
                     GuiceBundle.builder()
-                            .bundleLookup(new VoidBundleLookup())
                             .enableAutoConfig(FooResource.package.name)
                             .searchCommands()
-                            .bundles(
-                            new FooBundle(),
-                            new GuiceRestrictedConfigBundle())
+                            .bundles(new FooBundle())
                             .modules(new FooModule())
                             .disableInstallers(LifeCycleInstaller)
                             .strictScopeControl()